move Cube and ThreeControls into World, also some styling of Panel
