+++ rotating-cubes/V2.0 react/src/Demo/World/Cube/helpers/getAxisFromText.tsx
@@ -1,5 +1,5 @@
 import { Vector3 } from 'three';
-import { axisType } from '../../Types/types';
+import { axisType } from '../../../Types/types';
 
 export const getAxisFromText = (finalAxis: axisType) => {
     switch (finalAxis) {
